Fix 64-bit warnings in BVT

+++ mpich2/test/mpi/attr/attrend2.c
@@ -44,7 +44,7 @@ int main(int argc, char **argv)
         /* create the keyval for the exit handler */
         MPI_Comm_create_keyval(MPI_COMM_NULL_COPY_FN, delete_fn, &exit_keys[i], NULL);
         /* attach to comm_self */
-        MPI_Comm_set_attr(MPI_COMM_SELF, exit_keys[i], (void*)i);
+        MPI_Comm_set_attr(MPI_COMM_SELF, exit_keys[i], (void*)(size_t)i);
     }
 
     /* we can free the keys now */
@@ -93,7 +93,7 @@ int delete_fn(MPI_Comm comm, int keyval, void *attribute_val, void *extra_state)
 {
     int flag;
     int i;
-    int my_idx = (int)attribute_val;
+    int my_idx = (int)(size_t)attribute_val;
 
     if (my_idx < 0 || my_idx > NUM_TEST_ATTRS) {
         printf("internal error, my_idx=%d is invalid!\n", my_idx);
